introduce ai crate with completion providers
