+++ crates/ai/src/ai.rs
@@ -1,0 +1,1 @@
+pub mod completion;
